backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.3
python-dotenv==1.0.1
pydantic==2.9.2
bcrypt==4.2.1  # or latest compatible version
python-jose[cryptography]==3.3.0